[backend] Refactor upsert
